+++ src/Field.php
@@ -101,10 +101,10 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
     public $maxTopBlocks;
 
     /**
-     * @var int|null The maximum level that blocks in this field can be nested.
+     * @var int|null The maximum number of levels that blocks in this field can be nested.
      * @since 2.9.0
      */
-    public $maxLevel;
+    public $maxLevels;
 
     public $wasModified = false;
 
@@ -183,7 +183,7 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
                 self::PROPAGATION_METHOD_ALL
             ]
         ];
-        $rules[] = [['minBlocks', 'maxBlocks', 'maxTopBlocks', 'maxLevel'], 'integer', 'min' => 0];
+        $rules[] = [['minBlocks', 'maxBlocks', 'maxTopBlocks', 'maxLevels'], 'integer', 'min' => 0];
 
         return $rules;
     }
@@ -478,7 +478,7 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
             [
                 FieldValidator::class,
                 'maxTopBlocks' => $this->maxTopBlocks ?: null,
-                'maxLevel' => $this->maxLevel ?: null,
+                'maxLevels' => $this->maxLevels ?: null,
                 'on' => Element::SCENARIO_LIVE,
             ],
         ];