+++ src/services/Blocks.php
@@ -199,7 +199,7 @@ class Blocks extends Component
     {
         $dbService = Craft::$app->getDb();
         $structuresService = Craft::$app->getStructures();
-        $fieldMaxLevel = Craft::$app->getFields()->getFieldById($blockStructure->fieldId)->maxLevel ?: null;
+        $fieldMaxLevels = Craft::$app->getFields()->getFieldById($blockStructure->fieldId)->maxLevel ?: null;
         $record = new BlockStructureRecord();
 
         $transaction = $dbService->beginTransaction();
@@ -210,11 +210,11 @@ class Blocks extends Component
 
             if (!$structure) {
                 $structure = new Structure();
-                $structure->maxLevels = $fieldMaxLevel;
+                $structure->maxLevels = $fieldMaxLevels;
                 $structuresService->saveStructure($structure);
                 $blockStructure->structureId = $structure->id;
-            } else if ($structure->maxLevels !== $fieldMaxLevel) {
-                $structure->maxLevels = $fieldMaxLevel;
+            } else if ($structure->maxLevels !== $fieldMaxLevels) {
+                $structure->maxLevels = $fieldMaxLevels;
                 $structuresService->saveStructure($structure);
             }